+++ packages/plugins/m-function-matcher/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inlang/plugin-m-function-matcher
 
+## 0.9.20
+
+### Patch Changes
+
+- Updated dependencies [d27a983]
+- Updated dependencies [a27b7a4]
+  - @inlang/sdk@0.34.10
+
 ## 0.9.19
 
 ### Patch Changes